requirements.txt
supabase=2.15.1
discord.py-interactions[voice]=5.14.0